orm_sqlfan/orm_sqlfan/requirements.txt
backcall==0.1.0
decorator==4.4.1
Django==2.2.7
django-extensions==2.2.5
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
djangorestframework==3.12.2
django-filter==2.4.0
Pillow==8.2.0
drf-extra-fields==3.1.1
djoser==2.1.0
django-rest-knox==4.1.0